// BackEnd/src/schemas/status.enum.ts
export enum Status {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  DELETED = 'DELETED',
  SUSPENDED = 'SUSPENDED',
}

// Enum for voucher status
export enum VoucherStatus {
  EMAIL_SENT = 'EMAIL_SENT',
  CUSTOMER_FILLED_CHOICES = 'CUSTOMER_FILLED_CHOICES',
  MODIFIED = 'MODIFIED',
  FINAL_CHOICE = 'FINAL_CHOICE',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
}